fix: refacto les indicateurs deca
