+++ server/src/common/actions/indicateurs/indicateurs-national.actions.ts
@@ -5,7 +5,8 @@ import { tryCachedExecution } from "@/common/utils/cacheUtils";
 
 import { DateFilters, TerritoireFilters } from "../helpers/filters";
 
-import { getIndicateursEffectifsParDepartement, getIndicateursOrganismesParDepartement } from "./indicateurs.actions";
+import { getIndicateursEffectifsParDepartement } from "./indicateurs-with-deca.actions";
+import { getIndicateursOrganismesParDepartement } from "./indicateurs.actions";
 
 const indicateursNationalCacheExpirationMs = 3600 * 1000; // 1 hour
